making progress on bug
adding stub for input namelist for Const_8
I don't think it actually needs a full namelist in the original flora
but it includes dowriteb which is a flag for printing debug statements
while calculating magnetic field in equiltm.

Now the logs diverge in a new way. It seems to get to 50th omegaxb step.