adding GLREXE loop to increment time in main.F
There was a bug in rightvec.F
The 9-point spatial derivative stencil
has array index issue when sampling on radial boundary layers.

Temporarily removed those samples. Not by setting the physical quantity to 0, but by
setting the array index offset to 0.

There is a new bug where initial perturbation is 0, and hence does not
grow. This might be a simple input issue.