flora now saves the last frame of XRO to netcdf file.
The next step is to make the CDF writer more sophisticated.
Instead of a single call at end of program (glrfin).

A CDFinit declaring variables and dimensions at the beginning.
A CDFwrite which can be called at each time step as necessary.
A CDFfinish to close the file at end of program.