FLORA is now writing the XRO array,
saved at each time step, to the netcdf output file.
There was an issue with (start,count) in nf90_put_var
so I created a container variable that is a 2D array,
one column of which is written each time step.
And the entire 2D array is written ONCE in glrfin.